organized functions and added documentation
